@@ -7,8 +7,8 @@
   },
   "dependencies": {
     "lodash": "^4.13.1",
-    "react": "15.0.2",
-    "react-native": "^0.26.3",
+    "react": "^15.2.0",
+    "react-native": "^0.28.0",
     "react-native-video": "^0.8.0"
   }
 }
