@@ -9,8 +9,8 @@
     "lodash": "^3.10.1",
     "md5": "^2.1.0",
     "moment": "^2.13.0",
-    "react": "15.0.2",
-    "react-native": "^0.26.3",
+    "react": "^15.2.0",
+    "react-native": "^0.28.0",
     "reflux": "^0.4.1"
   }
 }
