@@ -1,25 +1,31 @@
-var React = require('react-native');
-var { SwitchIOS } = React;
+import React, {
+  Component,
+} from 'react';
 
-var Switch = React.createClass({
-  getInitialState() {
-    return {value: false};
-  },
+import {
+  SwitchIOS,
+} from 'react-native';
+
+class Switch extends Component{
+  constructor(props) {
+    super(props);
+    this.state = {value: false};
+  }
 
   _onValueChange(value) {
     this.setState({value: value});
     if (this.props.onValueChange) {
       this.props.onValueChange(value);      
     }
-  },
+  }
 
   render() {
     return (
       <SwitchIOS
-        onValueChange={this._onValueChange}
+        onValueChange={this._onValueChange.bind(this)}
         value={this.state.value}/>
       );
   }
-});
+}
 
-module.exports = Switch;
+export default Switch;
