@@ -27,4 +27,4 @@ class Card {
   }
 }
 
-module.exports = Card;
+export default Card;
