@@ -27,4 +27,4 @@ class Deck {
   }
 }
 
-module.exports = Deck;
+export default Deck;
