@@ -68,4 +68,4 @@ class Review {
   }
 }
 
-module.exports = Review;
+export default Review;
