@@ -3,13 +3,13 @@
  * https://github.com/facebook/react-native
  */
 
-import React, {
-  AppRegistry,
-  Component,
-  StyleSheet,
-  Text,
-  View
-} from 'react-native';
+ import React, { Component } from 'react';
+ import {
+     AppRegistry,
+     StyleSheet,
+     Text,
+     View
+ } from 'react-native';
 
 class Project extends Component {
   render() {
