@@ -102,7 +102,6 @@ const styles = StyleSheet.create({
     padding: 30
   },
   zipContainer: {
-    flex: 1,
     borderBottomColor: '#DDDDDD',
     borderBottomWidth: 1,
     marginLeft: 5,
@@ -111,13 +110,12 @@ const styles = StyleSheet.create({
   zipCode: {
     width: 50,
     height: baseFontSize,
+    padding: 0
   },
   mainText: {
-    flex: 1,
     fontSize: baseFontSize,
     color: '#FFFFFF'
   }
 });
 
 export default WeatherProject;
-
